+++ config/core/autopairs.nix
@@ -1,9 +1,0 @@
-{ pkgs, ... }:
-{
-  vim.extraPlugins = with pkgs.vimPlugins; {
-    nvim-autopairs = {
-      package = nvim-autopairs;
-      setup = ''require("nvim-autopairs").setup {}'';
-    };
-  };
-}