+++ config/plugins/floaterm.nix
@@ -1,8 +1,0 @@
-{ pkgs, ... }:
-{
-  vim.extraPlugins = {
-    floaterm = {
-      package = pkgs.vimPlugins.vim-floaterm;
-    };
-  };
-}