+++ config/plugins/mini-nvim.nix
@@ -4,6 +4,9 @@
     startPlugins = [
       pkgs.vimPlugins.mini-nvim
     ];
+    # mini.pairs is the only autopair provider. nvim-autopairs used to be set
+    # up alongside it in core/autopairs.nix, so both plugins inserted closing
+    # characters on the same keystroke.
     pluginRC.mini-pairs = ''
       require('mini.pairs').setup()
     '';