+++ config/utility/snacks-nvim.nix
@@ -1,7 +1,0 @@
-{
-  vim.utility = {
-    snacks-nvim = {
-      enable = true;
-    };
-  };
-}